feat: manage listings & user auth improvements

+++ server/app/Http/Controllers/ListingController.php
@@ -43,13 +43,18 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
-        // Session::flash('message','Listing create');
+        
         return redirect('/')->with('message','Listing created succesfully!');
     }
 
     // update listing data
     public function update(Request $request, Listing $listing){
+        // Make sure logged in uder is owner
+        if($listing->user_id != auth()->id()) abort(403,'Unauthorized action');
+
         $formFields = $request->validate([
             'title'=>'required',
             'company'=>['required'],
@@ -76,8 +81,16 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing){
+        // Make sure logged in uder is owner
+        if($listing->user_id != auth()->id()) abort(403,'Unauthorized action');
+
         $listing->delete();
         return redirect('/')->with('message','Listing Deleted Succesfully!');
+    }
+
+    // manage listings
+    public function manage(){
+        return view('listings.manage',['listings' => auth()->user()->listings()->get()]);
     }
 }
